performance-api/performance-observer-no-document-leak.html is flaky https://bugs.webkit.org/show_bug.cgi?id=186938 <rdar://problem/41379336>
Reviewed by Simon Fraser.

Source/WebCore:

Add internals API to get the identifier of a document and to ask if the document with
a given identifier is still alive. This is helpful to write tests for document leaking
fixes.

* testing/Internals.cpp:
(WebCore::Internals::documentIdentifier const):
(WebCore::Internals::isDocumentAlive const):
* testing/Internals.h:
* testing/Internals.idl:

LayoutTests:

Update test to stop relying on internals.numberOfLiveDocuments() and instead rely on the new
internals.documentIdentifier() / internals.isDocumentAlive(documentIdentifier) API in order
to address the flakiness. Relying on the number of live documents to check if a particular
document was destroyed is unreliable and flaky given that WebKit constructs documents for
various reasons.

* TestExpectations:
* performance-api/performance-observer-no-document-leak-expected.txt:
* performance-api/performance-observer-no-document-leak.html:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@233490 268f45cc-cd09-0410-ab3c-d52691b4dbfc