REGRESSION: Web Inspector no longer pauses in internal injected scripts like WDFindNodes.js https://bugs.webkit.org/show_bug.cgi?id=187350 <rdar://problem/41728249>
Reviewed by Matt Baker.

Source/JavaScriptCore:

Add a new command that toggles whether or not to blackbox internal scripts.
If blackboxed, the scripts will not be shown to the frontend and the debugger will
not pause in source frames from blackboxed scripts. Sometimes we want to break into
those scripts when debugging Web Inspector, WebDriver, or other WebKit-internal code
that injects scripts.

* inspector/agents/InspectorDebuggerAgent.cpp:
(Inspector::InspectorDebuggerAgent::setPauseForInternalScripts):
(Inspector::InspectorDebuggerAgent::didParseSource):
* inspector/agents/InspectorDebuggerAgent.h:
* inspector/protocol/Debugger.json:

Source/WebInspectorUI:

* UserInterface/Base/Setting.js: Add a new setting to allow pausing in internal scripts.
* UserInterface/Controllers/DebuggerManager.js: Listen to the setting change and toggle
the backend setting accordingly. The default is to not break into internal scripts.

* UserInterface/Views/SettingsTabContentView.js:
(WI.SettingsTabContentView.prototype._createDebugSettingsView):
Expose the new setting in the Debug settings panel.

LayoutTests:

Add a new test to demonstrate a difference in behavior when Debugger.setPauseForInternalScripts(true)
is sent to the backend. There's no way to correctly stringify the current call frame when stepping into
console.log because the injected script was blackboxed at the time that the scripts were parsed.
Instead, the stepping output just shows that the debugger is paused somewhere inside console.log.

* inspector/debugger/pause-for-internal-scripts-expected.txt: Added.
* inspector/debugger/pause-for-internal-scripts.html: Added.
* inspector/debugger/resources/log-pause-location.js:
(TestPage.registerInitializer.window.addSteppingTestCase):
(TestPage.registerInitializer):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@233655 268f45cc-cd09-0410-ab3c-d52691b4dbfc