[iOS] [WK1] Crash inside IOSurfacePool::platformGarbageCollectNow() in WebThread https://bugs.webkit.org/show_bug.cgi?id=187635 <rdar://problem/34297065>
Reviewed by Simon Fraser.

r167717 added code to trigger a CA commit in the web process via platformGarbageCollectNow() in order to free IOSurface-related memory.
However, that code is also running in the web thread in apps using WebKit1, causing unwanted UIView layout on the web thread.

Fix by not triggering this CA commit if it's called on the web thread.

* platform/graphics/cocoa/IOSurfacePoolCocoa.mm:
(WebCore::IOSurfacePool::platformGarbageCollectNow):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@233816 268f45cc-cd09-0410-ab3c-d52691b4dbfc