[WPE] Update WPEBackend in flatpak https://bugs.webkit.org/show_bug.cgi?id=187753
r233763 updated WPEBackend for the jhbuild, but not the flatpak. This
caused WPE_BACKEND_CHECK_VERSION to not be defined, triggering a bug
in the version guard,

#if defined(WPE_BACKEND_CHECK_VERSION) && WPE_BACKEND_CHECK_VERSION(0, 2, 0)

This doesn't work as intended, since the C preprocessor first expands
all macro arguments in the #if expression before evaluating the
expression. When WPE_BACKEND_CHECK_VERSION is not defined, this will
lead to a syntax error and some head scratching.

A future patch should address the buggy macro check, when it is
decided whether the assume the macro is always defined and bump
the API requirements on WPEBackend, or to move the check into two
conditionals to avoid the expansion problem.

Unreviewed build fix.


* flatpak/org.webkit.WPE.yaml:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@233914 268f45cc-cd09-0410-ab3c-d52691b4dbfc