REGRESSION(r234055): inverted return values in MediaPlayer::wouldTaintOrigin() https://bugs.webkit.org/show_bug.cgi?id=187944
Reviewed by Carlos Garcia Campos.

The MediaPlayer::wouldTaintOrigin(), as introduced in r234055, should
return false when the URL protocol is `data:`, and should return true
when the given SecurityOrigin cannot access the media URL.

This matches behavior of CanvasRenderingContext::wouldTaintOrigin()
which the new method falls back to when the MediaPlayerPrivate
implementation doesn't or can't determine the would-taint result.

This fixes a bunch of failing and timing-out WebGL test cases on ports
using GStreamer (on which the MediaPlayerPrivate derivate doesn't
implement the wouldTaintOrigin() method).

* platform/graphics/MediaPlayer.cpp:
(WebCore::MediaPlayer::wouldTaintOrigin const):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@234140 268f45cc-cd09-0410-ab3c-d52691b4dbfc