[LFC][BFC] Special case the document renderer when computing the height and margin https://bugs.webkit.org/show_bug.cgi?id=188313
Reviewed by Antti Koivisto.

The document renderer is certainly special and its geometry is likely to be computed as part of "10.6.6 Complicated cases" (spec is unclear) but probably not by
making the it a formatting root. Let's just special case it at contentHeightForFormattingContextRoot(), since having the document renderer as a context root has unwanted side effect.

* layout/FormattingContextGeometry.cpp:
(WebCore::Layout::contentHeightForFormattingContextRoot):
* layout/layouttree/LayoutBox.cpp:
(WebCore::Layout::Box::establishesBlockFormattingContext const):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@234554 268f45cc-cd09-0410-ab3c-d52691b4dbfc