navigator.sendBeacon does not work in pagehide callbacks https://bugs.webkit.org/show_bug.cgi?id=188329
Reviewed by Alex Christensen.

Source/WebCore:

Add support for sending beacons from pagehide event handlers. We normally do not allow loads because we're
about to enter PageCache. However, in case of Beacon, this is fine since it uses PingLoad and does not
WebCore to do the load.

Test: http/wpt/beacon/sendBeacon-in-pagehide.html

* loader/cache/CachedResource.cpp:
(WebCore::CachedResource::load):
- Allow Beacon loads to go through even if the document's pageCacheState is AboutToEnterPageCache (i.e.
we're firing the 'pagehide' event)
- Allow Becon loads to go though even if the FrameLoader's state is provisional (i.e. a load is pending)

LayoutTests:

Add layout test coverage.

* http/wpt/beacon/sendBeacon-in-pagehide-expected.txt: Added.
* http/wpt/beacon/sendBeacon-in-pagehide.html: Added.
* http/wpt/beacon/support/sendBeacon-onpagehide-window.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@234684 268f45cc-cd09-0410-ab3c-d52691b4dbfc