[win-ews] No such file or directory: u'/cygdrive/C/cygwin/home/buildbot/WebKit/WebKitBuild/Release/bin32/layout-test-results/pywebsocket.ws.log-out.txt' https://bugs.webkit.org/show_bug.cgi?id=188640
Unreviewed test gardening.

For now, skip websocket tests on Windows. Currently, run-webkit-tests is failing to start the websocket server because
port 8880 is already in use. Port 8880 is being used because the python websocket server process from the previous run
did not shutdown properly. We still need to understand why this is happening, but disable websocket tests now in order
for tests to run again.

* platform/win/TestExpectations:



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@234933 268f45cc-cd09-0410-ab3c-d52691b4dbfc