Web Inspector: Color picker: can't enter decimal numbers for opacity https://bugs.webkit.org/show_bug.cgi?id=187026 <rdar://problem/41446500>
Reviewed by Brian Burg.

After every "input" event, we update the `color` value of the `WI.ColorPicker` based on a
generated string using the values of the various <input>. The issue with this approach is
that adding a decimal point (e.g. "0.") would still be construed as 0, meaning that the
color wouldn't change and would instead be reset back to it's old value. This patch adds an
early return if the newly generated color has the same value as the current color, thereby
meaning that the `color` wouldn't change when changing from "0" to "0.".

* UserInterface/Views/ColorPicker.js:
(WI.ColorPicker):
(WI.ColorPicker.createColorInput):
(WI.ColorPicker.prototype._handleColorInputInput):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@235273 268f45cc-cd09-0410-ab3c-d52691b4dbfc