Don't launch network process in WebCookieManagerProxy::setHTTPCookieAcceptPolicy https://bugs.webkit.org/show_bug.cgi?id=188906 <rdar://problem/43539661>
Reviewed by Geoffrey Garen.

In WebCookieManagerProxy::setHTTPCookieAcceptPolicy, we persist the cookieAcceptPolicy of
sharedCookieStorage. When we launch the network process later, we pass the
identifier of sharedCookieStorage to network process, so network process has the correct
cookieAcceptPolicy. Therefore, we don't have to launch the network process and send the
setting message if the network process is not launched.

* UIProcess/WebCookieManagerProxy.cpp:
(WebKit::WebCookieManagerProxy::setHTTPCookieAcceptPolicy):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@235321 268f45cc-cd09-0410-ab3c-d52691b4dbfc