[CMake] Allow port specific options on gtest https://bugs.webkit.org/show_bug.cgi?id=189313
Reviewed by Alex Christensen.

Source/ThirdParty:

Allows a port to configure options within gtest for its platform.
The library type can be set otherwise it defaults to shared. Also the
compile definitions are propogated to TestWebKitAPI through a global
property.

* gtest/CMakeLists.txt:
* gtest/PlatformWin.cmake: Added.

Tools:

Use the global property set by the gtest CMake file to add to the
compilation definitions for TestWebKitAPI rather than repeating the
logic.

* TestWebKitAPI/CMakeLists.txt:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@235718 268f45cc-cd09-0410-ab3c-d52691b4dbfc